Changes in description
- Version bump to 2.0.0
- Added more API related functions for external plugins
- Moved files to the SalmonDE\TopVoter namespace to match the plugin
guidelines of pmmp more
- Moved events to an EventListener for more tidiness (There's still none
;-;)
- Possibly prevented a bug
