+++ plugin.yml
@@ -1,5 +1,5 @@
 name: TopVoter
-version: 1.2.0
+version: 2.0.0
 api: [2.0.0, 3.0.0, 3.0.0-ALPHA1, 3.0.0-ALPHA2, 3.0.0-ALPHA3]
 main: SalmonDE\TopVoter
 author: SalmonDE
